+++ src/Runner.js
@@ -1,12 +1,14 @@
 import EventEmitter from 'events';
 
 const stepsSymbol = Symbol( 'steps' );
+const stepResultsSymbol = Symbol( 'stepResults' );
 
 class Runner extends EventEmitter {
 	constructor() {
 		super();
 
 		this[ stepsSymbol ] = Object.freeze( new Set() );
+		this[ stepResultsSymbol ] = {};
 	}
 
 	get steps() {
@@ -14,7 +16,11 @@ class Runner extends EventEmitter {
 	}
 
 	addStep( step ) {
-		if ( !isValidStep( step ) ) {
+		const stepIds = [ ...this.steps ].map( ( { id } ) => {
+			return id;
+		} );
+
+		if ( !isValidStep( step, stepIds ) ) {
 			throw new TypeError( 'Provided object must be a valid step definition' );
 		}
 
@@ -54,21 +60,26 @@ class Runner extends EventEmitter {
 			return result;
 		};
 		const step = steps.shift();
+		const context = this._constructContext( path );
 
 		if ( !step ) {
 			return finish( true );
 		}
 
-		this.emit( 'step:start', step );
+		this.emit( 'step:start', step, context );
 
 		try {
-			const result = await step.run( path );
+			const requiresParameter = this._constructRequiresParameter( step );
+
+			const result = await step.run( path, requiresParameter );
 
 			if ( !isValidResult( result ) ) {
 				throw new TypeError( `Step ${ step.name } didn't return correct results` );
 			}
 
-			this.emit( 'step:end', step, result );
+			this[ stepResultsSymbol ][ step.id ] = result;
+
+			this.emit( 'step:end', step, result, context );
 
 			if ( !result.ok ) {
 				return finish( false );
@@ -85,13 +96,33 @@ class Runner extends EventEmitter {
 
 		return this._processSteps( steps, path );
 	}
+
+	_constructRequiresParameter( { requires } ) {
+		if ( !requires ) {
+			return {};
+		}
+
+		return requires.reduce( ( parameter, stepId ) => {
+			const results = {
+				[ stepId ]: this[ stepResultsSymbol ][ stepId ]
+			};
+
+			return { ...parameter, ...results };
+		}, {} );
+	}
+
+	_constructContext( projectPath ) {
+		return {
+			projectPath
+		};
+	}
 }
 
 function isNonEmptyString( value ) {
 	return typeof value === 'string' && value.trim().length > 0;
 }
 
-function isValidStep( step ) {
+function isValidStep( step, stepIds ) {
 	if ( !step || typeof step !== 'object' ) {
 		return false;
 	}
@@ -99,8 +130,10 @@ function isValidStep( step ) {
 	const isIdValid = isValidStepId( step.id );
 	const isNameValid = typeof step.name === 'string' && step.name.trim().length > 0;
 	const isRunValid = typeof step.run === 'function';
+	const isReporterValid = typeof step.report === 'function';
+	const isRequiresValid = Array.isArray( stepIds ) ? isValidRequires( step.requires, stepIds ) : true;
 
-	return isIdValid && isNameValid && isRunValid;
+	return isIdValid && isNameValid && isRunValid && isReporterValid && isRequiresValid;
 
 	function isValidStepId( id ) {
 		if ( typeof id !== 'string' || id.trim().length === 0 ) {
@@ -113,6 +146,20 @@ function isValidStep( step ) {
 
 		return isLoweredCase && isSpaceless;
 	}
+
+	function isValidRequires( requires, availableSteps ) {
+		if ( typeof requires === 'undefined' ) {
+			return true;
+		}
+
+		if ( typeof requires !== 'undefined' && !Array.isArray( requires ) ) {
+			return false;
+		}
+
+		return requires.every( ( requires ) => {
+			return availableSteps.includes( requires );
+		} );
+	}
 }
 
 function isValidResult( stepResults ) {
@@ -121,8 +168,8 @@ function isValidResult( stepResults ) {
 	}
 
 	const isValidResults = stepResults.results && typeof stepResults.results === 'object';
-	const isValidReporter = typeof stepResults.reporter === 'function';
-	return isValidResults && isValidReporter;
+
+	return isValidResults;
 }
 
 export default Runner;