+++ src/index.js
@@ -1,3 +1,11 @@
 import mlt from './mlt.js';
+import linter from './linter.js';
+import tester from './tester.js';
+import codeCoverage from './codeCoverage.js';
+import codecov from './codecov.js';
 
-export default mlt;
+export { mlt };
+export { linter };
+export { tester };
+export { codeCoverage };
+export { codecov };