+++ src/steps.js
@@ -1,9 +1,12 @@
 /* istanbul ignore file */
 
-import linter from './linter.js';
-import tester from './tester.js';
+import { spawn } from 'threads';
+import { Thread } from 'threads';
+import { Worker } from 'threads';
 import codeCoverage from './codeCoverage.js';
-import codecov from './codecov.js';
+import linterReporter from './reporters/linter';
+import codeCoverageReporter from './reporters/codeCoverage';
+import codecovReporter from './reporters/codecov';
 
 const steps = [
 	{
@@ -11,8 +14,9 @@ const steps = [
 		name: 'Linter',
 		watchable: true,
 		run( projectPath ) {
-			return linter( projectPath );
-		}
+			return spawnWorker( './workers/linter.js', projectPath );
+		},
+		report: linterReporter
 	},
 
 	{
@@ -20,7 +24,10 @@ const steps = [
 		name: 'Tester',
 		watchable: true,
 		run( projectPath ) {
-			return tester( projectPath );
+			return spawnWorker( './workers/tester.js', projectPath );
+		},
+		report( { output }, logger ) {
+			logger.log( output );
 		}
 	},
 
@@ -28,9 +35,13 @@ const steps = [
 		id: 'coverage',
 		name: 'Code Coverage',
 		watchable: true,
-		run( projectPath ) {
-			return codeCoverage( projectPath, global.__mltCoverage__ );
-		}
+		requires: [
+			'test'
+		],
+		run( projectPath, { test: { coverage } } ) {
+			return codeCoverage( projectPath, coverage );
+		},
+		report: codeCoverageReporter
 	},
 
 	{
@@ -38,9 +49,20 @@ const steps = [
 		name: 'CodeCov',
 		watchable: false,
 		run( projectPath ) {
-			return codecov( projectPath );
-		}
+			return spawnWorker( './workers/codecov.js', projectPath );
+		},
+		report: codecovReporter
 	}
 ];
+
+async function spawnWorker( workerPath, ...workerArgs ) {
+	const worker = new Worker( workerPath );
+	const workerAPI = await spawn( worker );
+	const result = await workerAPI( ...workerArgs );
+
+	await Thread.terminate( workerAPI );
+
+	return result;
+}
 
 export default steps;