+++ tests/index.js
@@ -1,7 +1,16 @@
-import mlt from '../src/index.js';
+import * as pkg from '../src/index.js';
 
 describe( 'index', () => {
 	it( 'exports a function', () => {
-		expect( mlt ).to.be.a( 'function' );
+		const expectedExports = [
+			'mlt',
+			'linter',
+			'tester',
+			'codeCoverage',
+			'codecov'
+		];
+		const actualExports = Object.keys( pkg );
+
+		expect( actualExports ).to.deep.equal( expectedExports );
 	} );
 } );